add Cube, CubesScene. Fix zbuffer. Change to double

+++ DirectX Engine/Camera.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Vec3f.h"
+#include "Vec3.h"
 #include "Matrix3.h"
 #include "MathHelper.h"
 #include "Vec2f.h"
@@ -7,24 +7,24 @@
 class Camera
 {
 public:
-	Vec3f pos = Vec3f();
-	Vec3f X90 = Vec3f(1.0f, 0.0f, 0.0f);
-	Vec3f Y90 = Vec3f(0.0f, 1.0f, 0.0f);
-	Vec3f Z90 = Vec3f(0.0f, 0.0f, 1.0f);
+	Vec3d pos = Vec3d();
+	Vec3d X90 = Vec3d(1.0, 0.0, 0.0);
+	Vec3d Y90 = Vec3d(0.0, 1.0, 0.0);
+	Vec3d Z90 = Vec3d(0.0, 0.0, 1.0);
 	Matrix3f transform;
-	Vec3f leftNormal;
-	Vec3f rightNormal;
-	Vec3f topNormal;
-	Vec3f bottomNormal;
-	float widthHalf;
-	float heightHalf;
-	float fovW;
-	float fovH;
-	float scaleX;
-	float scaleY;
+	Vec3d leftNormal;
+	Vec3d rightNormal;
+	Vec3d topNormal;
+	Vec3d bottomNormal;
+	double widthHalf;
+	double heightHalf;
+	double fovW;
+	double fovH;
+	double scaleX;
+	double scaleY;
 public:
-	Camera(int width, int height, float fov);
-	void Rotate(float x, float y);
-	Vec3f ToScreen(const Vec3f& v) const;
+	Camera(int width, int height, double fov);
+	void Rotate(double x, double y);
+	Vec3d ToScreen(const Vec3d& v) const;
 };
